// frontend/src/app/models/detaljiFirme.ts
export class DetaljiFirme{
    idF: string;
    kor_ime: string;
    nazivP: string;
    kategorija: string;
    ocena: number;
    br_zvezdica: number;
    popust: number;
    komentari: Array<string>;
    relax: boolean;
    vrste_soba: Array<string>;
    opremljenost: Array<string>;
    soba: Array<string>;
    hrana: Array<string>;
    dodatno: string;
    slike: Array<string>;
    ziro_racuni: Array<string>;
    casovi_za: Array<string>;
    cena_casa: number;
    instruktori: Array<string>;
    oprema: boolean;
    cena_opreme: number;
    proizvodjaci: Array<string>;
    planine: Array<string>;
    rezervacije: Array<string>;

}